+++ spec/rubocop/cop/performance/reverse_each_spec.rb
@@ -10,11 +10,26 @@ RSpec.describe RuboCop::Cop::Performance::ReverseEach do
     RUBY
   end
 
+  it 'registers an offense when each_with_index is called on reverse' do
+    expect_offense(<<~RUBY)
+      [1, 2, 3].reverse.each_with_index { |e, idx| puts [idx, e] }
+                ^^^^^^^^^^^^^^^^^^^^^^^ Use `reverse_each.with_index` instead of `reverse.each_with_index`.
+    RUBY
+  end
+
   it 'registers an offense when each is called on reverse on a variable' do
     expect_offense(<<~RUBY)
       arr = [1, 2, 3]
       arr.reverse.each { |e| puts e }
           ^^^^^^^^^^^^ Use `reverse_each` instead of `reverse.each`.
+    RUBY
+  end
+
+  it 'registers an offense when each_with_index is called on reverse on a variable' do
+    expect_offense(<<~RUBY)
+      arr = [1, 2, 3]
+      arr.reverse.each_with_index { |e, idx| puts [idx, e] }
+          ^^^^^^^^^^^^^^^^^^^^^^^ Use `reverse_each.with_index` instead of `reverse.each_with_index`.
     RUBY
   end
 
@@ -26,6 +41,17 @@ RSpec.describe RuboCop::Cop::Performance::ReverseEach do
 
       arr.reverse.each { |e| puts e }
           ^^^^^^^^^^^^ Use `reverse_each` instead of `reverse.each`.
+    RUBY
+  end
+
+  it 'registers an offense when each_with_index is called on reverse on a method call' do
+    expect_offense(<<~RUBY)
+      def arr
+        [1, 2, 3]
+      end
+
+      arr.reverse.each_with_index { |e, idx| puts [idx, e] }
+          ^^^^^^^^^^^^^^^^^^^^^^^ Use `reverse_each.with_index` instead of `reverse.each_with_index`.
     RUBY
   end
 
@@ -42,6 +68,19 @@ RSpec.describe RuboCop::Cop::Performance::ReverseEach do
     RUBY
   end
 
+  it 'registers an offense for a multi-line reverse.each_with_index' do
+    expect_offense(<<~RUBY)
+      def arr
+        [1, 2, 3]
+      end
+
+      arr.
+        reverse.
+        ^^^^^^^^ Use `reverse_each.with_index` instead of `reverse.each_with_index`.
+        each_with_index { |e, idx| puts [idx, e] }
+    RUBY
+  end
+
   it 'does not register an offense when reverse is used without each' do
     expect_no_offenses('[1, 2, 3].reverse')
   end
@@ -50,11 +89,21 @@ RSpec.describe RuboCop::Cop::Performance::ReverseEach do
     expect_no_offenses('[1, 2, 3].each { |e| puts e }')
   end
 
+  it 'does not register an offense when each_with_index is used without reverse' do
+    expect_no_offenses('[1, 2, 3].each_with_index { |e, idx| puts [idx, e] }')
+  end
+
   context 'autocorrect' do
     it 'corrects reverse.each to reverse_each' do
       new_source = autocorrect_source('[1, 2].reverse.each { |e| puts e }')
 
       expect(new_source).to eq('[1, 2].reverse_each { |e| puts e }')
+    end
+
+    it 'corrects reverse.each_with_index to reverse_each.with_index' do
+      new_source = autocorrect_source('[1, 2].reverse.each_with_index { |e, idx| puts [idx, e] }')
+
+      expect(new_source).to eq('[1, 2].reverse_each.with_index { |e, idx| puts [idx, e] }')
     end
 
     it 'corrects reverse.each to reverse_each on a variable' do
@@ -66,6 +115,18 @@ RSpec.describe RuboCop::Cop::Performance::ReverseEach do
       expect(new_source).to eq(<<~RUBY)
         arr = [1, 2]
         arr.reverse_each { |e| puts e }
+      RUBY
+    end
+
+    it 'corrects reverse.each_with_index to reverse_each.with_index on a variable' do
+      new_source = autocorrect_source(<<~RUBY)
+        arr = [1, 2]
+        arr.reverse.each_with_index { |e, idx| puts [idx, e] }
+      RUBY
+
+      expect(new_source).to eq(<<~RUBY)
+        arr = [1, 2]
+        arr.reverse_each.with_index { |e, idx| puts [idx, e] }
       RUBY
     end
 
@@ -87,7 +148,25 @@ RSpec.describe RuboCop::Cop::Performance::ReverseEach do
       RUBY
     end
 
-    it 'corrects a multi-line reverse_each' do
+    it 'corrects reverse.each_with_index to reverse_each.with_index on a method call' do
+      new_source = autocorrect_source(<<~RUBY)
+        def arr
+          [1, 2]
+        end
+
+        arr.reverse.each_with_index { |e, idx| puts [idx, e] }
+      RUBY
+
+      expect(new_source).to eq(<<~RUBY)
+        def arr
+          [1, 2]
+        end
+
+        arr.reverse_each.with_index { |e, idx| puts [idx, e] }
+      RUBY
+    end
+
+    it 'corrects a multi-line reverse.each' do
       new_source = autocorrect_source(<<~RUBY)
         def arr
           [1, 2]
@@ -105,6 +184,27 @@ RSpec.describe RuboCop::Cop::Performance::ReverseEach do
 
         arr.
           reverse_each { |e| puts e }
+      RUBY
+    end
+
+    it 'corrects a multi-line reverse.each_with_index' do
+      new_source = autocorrect_source(<<~RUBY)
+        def arr
+          [1, 2]
+        end
+
+        arr.
+          reverse.
+          each_with_index { |e, idx| puts [idx, e] }
+      RUBY
+
+      expect(new_source).to eq(<<~RUBY)
+        def arr
+          [1, 2]
+        end
+
+        arr.
+          reverse_each.with_index { |e, idx| puts [idx, e] }
       RUBY
     end
   end